Finished refactor lexer.c lexer_get_tokens.c parser_get_commands.c + norme
